Small fix for previous commit.
Added support for querying client attribute state through `glGet*`.
All the `glGet*` calls will first see if the attribute passed in is an attribute recognized by `glEnableClientState`
and if so, will retrieve that from `GL.immediate.enabledClientAttributes`.
Otherwise, it will call into WebGL for the value of the attribute.
Modified test included.