Added an additional call: `SDL_SetEventHandler()` to SDL. This is not in the official API but it is needed for Mozilla Persona, which makes sure an event handler is on the stack when `navigator.id.request()` is called.
Using the standard SDL api, this can never be the case because an SDL
app retrieves events from SDL’s event queue independently of when
those events are added to the queue (through event handlers that SDL
itself registers).

With this additional call, apps can receive events directly in
response to an actual event handler which will still be on the stack
when the event is handled by the SDL-based application.